requirements.txt
smolagents
duckduckgo-search
requests
beautifulsoup4
python-dotenv
